FitGenie/backend/requirements.txt
Flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
pymongo==4.7.2
dnspython==2.6.1
flask-jwt-extended==4.6.0
certifi>=2024.2.2